tests/requirements-test.txt
pytest==8.0.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-benchmark==4.0.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
pytest-env==1.1.3
pytest-randomly==3.15.0
coverage==7.4.1
hypothesis==6.98.0
faker==22.5.0
requests-mock==1.11.0
responses==0.24.1
freezegun==1.4.0
pillow==10.2.0
numpy==1.26.3
shapely==2.0.2
rasterio==1.3.9
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
flask==3.0.0
flask-limiter==3.5.0
flasgger==0.9.7.1
werkzeug==3.0.1 